1) Added checking user when his create feedback for meeting 2) Added feature to show user's meetings history 3) Some fix in database 4) Added format of LocalDateTime in MeetingController

+++ TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/service/dto/MeetingDTO.java
@@ -19,6 +19,6 @@ public class MeetingDTO {
     private List<String> links;
     private LocalDateTime timeOfAction;
     private Integer ownerId;
-    private List<Long> confirmedUserIds;
-    private List<Long> wishingUserIds;
+    private List<Integer> confirmedUserIds;
+    private List<Integer> wishingUserIds;
 }
